Use composition instead of inheritance for Tile
Extending Rectangle was fun for achieving the dragging behaviour in an
elegant way, but rubbish for building on top of it further. Now that
we're actually move the Tile when dragging it, we can swap to using
composition which will make swapping to an image instead of a Rectangle
much simpler.